--- src/features/videoEditor/index.jsx
@@ -1,56 +1,0 @@
-import PropTypes from "prop-types";
-
-import Button from "@/shared/components/Button";
-import Modal from "@/shared/components/Modal";
-
-import TrimSlider from "./components/TrimSlider";
-import VideoPlayer from "./components/VideoPlayer";
-import useVideoEditor from "./hooks/useVideoEditor";
-
-const VideoEditor = ({ videoSrc }) => {
-  const {
-    trim,
-    duration,
-    playing,
-    playerRef,
-    error,
-    setError,
-    handlePlay,
-    handlePause,
-    handleDuration,
-    handleTrimChange,
-    handleProgress,
-    handleEdit,
-  } = useVideoEditor(videoSrc);
-
-  const closeModal = () => {
-    setError(null);
-  };
-
-  return (
-    <>
-      <VideoPlayer
-        ref={playerRef}
-        url={videoSrc}
-        playing={playing}
-        onDuration={handleDuration}
-        onProgress={handleProgress}
-        onPlay={handlePlay}
-        onPause={handlePause}
-      />
-      <TrimSlider trim={trim} duration={duration} onChange={handleTrimChange} />
-      <Button onClick={handleEdit}>편집 요청</Button>
-      {error && (
-        <Modal onClick={closeModal} buttonText="닫기">
-          <p className="text-red-600 mb-4">{error}</p>
-        </Modal>
-      )}
-    </>
-  );
-};
-
-VideoEditor.propTypes = {
-  videoSrc: PropTypes.string.isRequired,
-};
-
-export default VideoEditor;
